facebook
A Facebook login page crafted using HTML and CSS focuses on creating a user-friendly and visually appealing interface for users to access their accounts. HTML provides the structural elements such as input fields for the username and password, buttons for logging in and signing up, and links for password recovery. CSS enhances the aesthetic and usability of the page by styling these components with appropriate colors, fonts, and layouts to align with Facebook's branding. This combination ensures a seamless and consistent user experience, making the login process straightforward and visually coherent with the rest of Facebook's platform.